fix: circular dependency in swap package
